refactor: moved search logic to operator and new way to specify search type
