+++ search/src/pages/collection/[collectionID].astro
@@ -16,7 +16,7 @@ const requestParams = Astro.request.url.split("?")[1];
 const params = new URLSearchParams(requestParams);
 const page = Number(params.get("page")) || 1;
 const query = params.get("q") ?? "";
-const searchType: string = params.get("searchType") ?? "search";
+const searchType: string = params.get("searchType") ?? "semantic";
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 let metadataFilters: any[] = [];
@@ -110,6 +110,9 @@ try {
           // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
           filters: dataTypeFilters.metadataFilters,
           collection_id: collectionID,
+          search_type: {
+            search_type: searchType,
+          }
         }),
       },
     );
