+++ search/src/pages/search.astro
@@ -13,7 +13,7 @@ const requestParams = Astro.request.url.split("?")[1];
 const params = new URLSearchParams(requestParams);
 const query = params.get("q") ?? "";
 const page = Number(params.get("page")) || 1;
-const searchType: string = params.get("searchType") ?? "search";
+const searchType: string = params.get("searchType") ?? "semantic";
 const weight = params.get("weight") ?? "0.5";
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -72,6 +72,9 @@ try {
         : null,
     // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
     filters: dataTypeFilters.metadataFilters,
+    search_type: {
+      searchType: searchType,
+    }
   };
 
   if (searchType === "hybrid_search") {
