+++ server/src/operators/mod.rs
@@ -6,6 +6,7 @@ pub mod message_operator;
 pub mod notification_operator;
 pub mod password_reset_operator;
 pub mod qdrant_operator;
+pub mod search_operator;
 pub mod topic_operator;
 pub mod user_operator;
 pub mod vote_operator;
